# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API para gestão integral da produção gráfica.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Solicitar URL segura para carregar ficheiro
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadUrlRequest" }
      responses:
        "200":
          description: URL de carregamento e caminho do ficheiro.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadUrlResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
...
  responses:
    BadRequest:
      description: Dados inválidos ou transição não permitida.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata: { $ref: "#/components/schemas/UploadUrlRequest" }